# swagger/Docs.yaml
openapi: 3.0.0
info:
  title: Team 695 API
  description: Team 695 API Documentation for multiple services including authentication, events, surveys, teams, uploads, and feedback
  version: 1.5.2
servers:
  - url: https://api.team695.com
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Also called Access Key
    apiSecretAuth:
      type: apiKey
      in: header
      name: X-API-Secret
      description: Used together with X-API-Key, also called Access Secret Key
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Optional message providing additional information
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error details (only in non-production environments)
...
    EndpointInfo:
      type: object
      properties:
        path:
          type: string
          description: API endpoint path
        method:
          type: string
          description: HTTP method
        summary:
          type: string
          description: Brief description of the endpoint
        description:
          type: string
          description: Detailed description of the endpoint
        tags:
          type: array
          items:
            type: string
          description: Tags associated with the endpoint
        security:
          type: array
          items:
            type: string
          description: Security requirements for the endpoint
        parameters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              in:
                type: string
              required:
                type: boolean
              description:
                type: string
        requestBody:
          type: object
          properties:
            required:
              type: boolean
            contentTypes:
              type: array
              items:
                type: string
        responses:
          type: array
          items:
            type: string
...
paths:
...
  /info/endpoints:
    get:
      summary: Get all available API endpoints
      description: Retrieve a list of all available API endpoints with their details
      tags:
        - Info
      responses:
        '200':
          description: Endpoints retrieved successfully
          headers:
            ETag:
              description: ETag of current swagger structure
              schema:
                type: string
            Last-Modified:
              description: Last modified timestamp of swagger structure
              schema:
                type: string
            Cache-Control:
              description: HTTP cache control policy
              schema:
                type: string
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          total:
                            type: integer
                          endpoints:
                            type: array
                            items:
                              $ref: '#/components/schemas/EndpointInfo'
                          tags:
                            type: array
                            items:
                              type: string
                          lastUpdated:
                            type: string
                            format: date-time
        '304':
          description: Not modified, use cached response
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /info/endpoints/by-tag/{tag}:
    get:
      summary: Get API endpoints by tag
      description: Retrieve API endpoints filtered by a specific Swagger tag
      tags:
        - Info
      parameters:
        - in: path
          name: tag
          required: true
          schema:
            type: string
          description: Tag name used to filter endpoints
      responses:
        '200':
          description: Endpoints for the tag retrieved successfully
          headers:
            ETag:
              description: ETag of current swagger structure
              schema:
                type: string
            Last-Modified:
              description: Last modified timestamp of swagger structure
              schema:
                type: string
            Cache-Control:
              description: HTTP cache control policy
              schema:
                type: string
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tag:
                            type: string
                          total:
                            type: integer
                          endpoints:
                            type: array
                            items:
                              $ref: '#/components/schemas/EndpointInfo'
                          lastUpdated:
                            type: string
                            format: date-time
        '304':
          description: Not modified, use cached response
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...